feat: refactor to v2.0.0
