--- .github/workflows/OFFgorelease.yaml.OFF
@@ -1,37 +1,0 @@
-name: 🎉 Release Binary
-
-# run only on tags
-on:
-  push:
-    tags:
-      - 'v*'
-
-permissions:
-   contents: write # needed to write releases
-   id-token: write # needed for keyless signing
-   packages: write # needed for ghcr access
-
-jobs:
-  release:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-
-      - name: Set up Go
-        uses: actions/setup-go@v3
-        with:
-          go-version: 1.19
-
-      - name: Run GoReleaser
-        uses: goreleaser/goreleaser-action@v3
-        with:
-          distribution: kvstok
-          version: latest
-          args: release --rm-dist
-          workdir: .
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
