--- .github/workflows/release-binary.yaml
@@ -1,27 +1,0 @@
-name: 🎉 Release Binary
-on:
-  push:
-    tags:
-      - '*'
-  workflow_dispatch:
-
-jobs:
-  release:
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-
-      - uses: actions/setup-go@v3
-        with:
-          go-version: 1.19
-
-      - uses: goreleaser/goreleaser-action@v3
-        with:
-          args: "release --rm-dist"
-          version: latest
-          workdir: .
-        env:
-          GITHUB_TOKEN: "${{ secrets.GITHUB_TOKEN }}"
