+++ .github/workflows/reviewdog.yaml
@@ -1,0 +1,23 @@
+name: 🐕 Review DOG
+on: [pull_request]
+jobs:
+  golangci-lint:
+    name: runner
+    runs-on: ubuntu-latest
+    steps:
+      - name: Check out code into the Go module directory
+        uses: actions/checkout@v3
+
+      - name: GolangCI Lint
+        uses: reviewdog/action-golangci-lint@v3
+
+      - name: Setup reviewdog
+        uses: reviewdog/action-setup@v1
+
+      - name: gofumpt -s with reviewdog
+        env:
+          REVIEWDOG_GITHUB_API_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          go install mvdan.cc/gofumpt@latest
+          gofumpt -e -d . | \
+          reviewdog -name="gofumpt" -f=diff -f.diff.strip=0 -reporter=github-pr-review
