+++ .github/workflows/typo-check.yaml
@@ -1,0 +1,12 @@
+name: 📬 Typo Check
+
+on: [push, pull_request]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: typos-action
+        uses: crate-ci/typos@latest
