--- OFF.goreleaser.yaml.OFF
@@ -1,69 +1,0 @@
-before:
-  hooks:
-    - go mod tidy
-
-snapshot:
-  name_template: '{{ incpatch .Version }}-next'
-
-builds:
-- env:
-    - CGO_ENABLED=0
-  goos: [ windows,linux,darwin ]
-  goarch: [ amd64,386,arm,arm64 ]
-  goarm: [ 7 ]
-  mod_timestamp: '{{ .CommitTimestamp }}'
-  flags:
-    - -trimpath
-  ldflags:
-    - -s -w -X main.version={{.Version}} -X main.commit={{.Commit}} -X main.date={{ .CommitDate }} -X main.builtBy=goreleaser
-
-universal_binaries:
-- replace: false
-
-checksum:
-  name_template: 'checksums.txt'
-
-changelog:
-  sort: asc
-  use: github
-  filters:
-    exclude:
-    - '^test:'
-    - '^chore'
-    - 'merge conflict'
-    - Merge pull request
-    - Merge remote-tracking branch
-    - Merge branch
-    - go mod tidy
-  groups:
-    - title: Dependency updates
-      regexp: "^.*(feat|fix)\\(deps\\)*:+.*$"
-      order: 300
-    - title: 'New Features'
-      regexp: "^.*feat[(\\w)]*:+.*$"
-      order: 100
-    - title: 'Bug fixes'
-      regexp: "^.*fix[(\\w)]*:+.*$"
-      order: 200
-    - title: 'Documentation updates'
-      regexp: "^.*docs[(\\w)]*:+.*$"
-      order: 400
-    - title: Other work
-      order: 9999
-
-archives:
-  - name_template: '{{ .ProjectName }}_{{ .Os }}_{{ .Arch }}{{ if .Arm }}v{{ .Arm }}{{ end }}'
-    replacements:
-      darwin: Darwin
-      linux: Linux
-      windows: Windows
-      386: i386
-      amd64: x86_64
-    format_overrides:
-    - goos: windows
-      format: zip
-    files:
-      - README.md
-      - LICENSE.md
-      - completions/*
-      - manpages/*
