+++ internal/entity/kventity.go
@@ -1,0 +1,7 @@
+package entity
+
+// Struct to store key/value
+type KVStok struct {
+	Key string
+	Val string
+}
