--- pkg/database/database.go
@@ -1,12 +1,0 @@
-package database
-
-import (
-	"github.com/xujiajun/nutsdb"
-)
-
-const (
-	DBName = ".6B7673" // -> .kvs
-	Bucket = "kvstok"
-)
-
-var DB *nutsdb.DB
