--- pkg/entity/kventity.go
@@ -1,6 +1,0 @@
-package entity
-
-type KVStok struct {
-	Key string
-	Val string
-}
